requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
Django==4.0.2
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
pytz==2021.3
requests==2.27.1
Rest-Framework-Auth-Toolkit==0.11
sqlparse==0.4.2
tomli==2.0.0
tzdata==2021.5
urllib3==1.26.8
